+++ test/extensions/persistent_session/plug/cookie_test.exs
@@ -5,8 +5,8 @@ defmodule PowPersistentSession.Plug.CookieTest do
   alias Plug.{Conn, ProcessStore, Test}
   alias Plug.Session, as: PlugSession
   alias Pow.{Plug, Plug.Session}
-  alias PowPersistentSession.{Plug.Cookie, Store.PersistentSessionCache}
-  alias PowPersistentSession.Test.Users.User
+  alias PowPersistentSession.{Plug.Cookie, Plug.Base, Store.PersistentSessionCache}
+  alias PowPersistentSession.Test.{Users.User, RepoMock}
   alias Pow.Test.EtsCacheMock
 
   @cookie_key "persistent_session"
@@ -18,8 +18,9 @@ defmodule PowPersistentSession.Plug.CookieTest do
 
     config = PowPersistentSession.Test.pow_config()
     conn   = conn_with_session_plug(config)
+    user   = RepoMock.get_by(User, [id: 1], [])
 
-    {:ok, conn: conn, config: config, ets: EtsCacheMock}
+    {:ok, conn: conn, config: config, user: user}
   end
 
   test "call/2 sets pow_persistent_session plug in conn", %{conn: conn, config: config} do
@@ -30,9 +31,8 @@ defmodule PowPersistentSession.Plug.CookieTest do
     refute conn.resp_cookies[@cookie_key]
   end
 
-  test "call/2 assigns user from cookie", %{conn: conn, ets: ets} do
-    user = %User{id: 1}
-    id   = store_in_cache(conn, "test", {[id: user.id], []}, backend: ets)
+  test "call/2 assigns user from cookie", %{conn: conn, user: user} do
+    id   = store_in_cache(conn, "test", {user, []})
     conn =
       conn
       |> persistent_cookie(@cookie_key, id)
@@ -41,8 +41,8 @@ defmodule PowPersistentSession.Plug.CookieTest do
     assert Plug.current_user(conn) == user
     assert %{value: new_id, max_age: @max_age, path: "/"} = conn.resp_cookies[@cookie_key]
     refute new_id == id
-    assert get_from_cache(conn, id, backend: ets) == :not_found
-    assert get_from_cache(conn, new_id, backend: ets) == {[id: 1], []}
+    assert get_from_cache(conn, id) == :not_found
+    assert get_from_cache(conn, new_id) == {user, []}
   end
 
   defmodule PersistentSessionCacheWaitDelete do
@@ -67,12 +67,11 @@ defmodule PowPersistentSession.Plug.CookieTest do
     end
   end
 
-  test "call/2 assigns user from cookie and doesn't expire with simultanous request", %{conn: conn, ets: ets} do
+  test "call/2 assigns user from cookie and doesn't expire with simultanous request", %{conn: conn, user: user} do
     :ets.delete(EtsCacheMock)
     :ets.new(EtsCacheMock, [:ordered_set, :public, :named_table])
 
-    user   = %User{id: 1}
-    id     = store_in_cache(conn, "test", {[id: user.id], []}, backend: ets)
+    id     = store_in_cache(conn, "test", {user, []})
     conn   = persistent_cookie(conn, @cookie_key, id)
     config = [persistent_session_store: {PersistentSessionCacheWaitDelete, ttl: :timer.hours(24) * 30, namespace: "persistent_session"}]
 
@@ -112,9 +111,8 @@ defmodule PowPersistentSession.Plug.CookieTest do
     task
   end
 
-  test "call/2 assigns user from cookie passing fingerprint to the session metadata", %{conn: conn, ets: ets} do
-    user = %User{id: 1}
-    id   = store_in_cache(conn, "test", {[id: user.id], session_metadata: [fingerprint: "fingerprint"]}, backend: ets)
+  test "call/2 assigns user from cookie passing fingerprint to the session metadata", %{conn: conn, user: user} do
+    id   = store_in_cache(conn, "test", {user, session_metadata: [fingerprint: "fingerprint"]})
     conn =
       conn
       |> persistent_cookie(@cookie_key, id)
@@ -123,14 +121,13 @@ defmodule PowPersistentSession.Plug.CookieTest do
     assert Plug.current_user(conn) == user
     assert %{value: new_id, max_age: @max_age, path: "/"} = conn.resp_cookies[@cookie_key]
     refute new_id == id
-    assert get_from_cache(conn, id, backend: ets) == :not_found
-    assert get_from_cache(conn, new_id, backend: ets) == {[id: 1], session_metadata: [fingerprint: "fingerprint"]}
+    assert get_from_cache(conn, id) == :not_found
+    assert get_from_cache(conn, new_id) == {user, session_metadata: [fingerprint: "fingerprint"]}
     assert conn.private[:pow_session_metadata][:fingerprint] == "fingerprint"
   end
 
-  test "call/2 assigns user from cookie passing custom metadata to session metadata", %{conn: conn, ets: ets} do
-    user = %User{id: 1}
-    id   = store_in_cache(conn, "test", {[id: user.id], session_metadata: [a: 1, b: 2, fingerprint: "fingerprint"]}, backend: ets)
+  test "call/2 assigns user from cookie passing custom metadata to session metadata", %{conn: conn, user: user} do
+    id   = store_in_cache(conn, "test", {user, session_metadata: [a: 1, b: 2, fingerprint: "fingerprint"]})
     conn =
       conn
       |> persistent_cookie(@cookie_key, id)
@@ -140,14 +137,13 @@ defmodule PowPersistentSession.Plug.CookieTest do
 
     assert Plug.current_user(conn) == user
     assert %{value: id, max_age: @max_age, path: "/"} = conn.resp_cookies[@cookie_key]
-    assert get_from_cache(conn, id, backend: ets) == {[id: 1], session_metadata: [fingerprint: "new_fingerprint", b: 2, a: 2]}
+    assert get_from_cache(conn, id) == {user, session_metadata: [fingerprint: "new_fingerprint", b: 2, a: 2]}
     assert [inserted_at: _, b: 2, a: 3, fingerprint: "new_fingerprint"] = conn.private[:pow_session_metadata]
   end
 
-  test "call/2 assigns user from cookie with prepended `:otp_app`", %{config: config, ets: ets} do
-    user = %User{id: 1}
+  test "call/2 assigns user from cookie with prepended `:otp_app`", %{config: config, user: user} do
     conn = conn_with_session_plug(config ++ [otp_app: :test_app])
-    id   = store_in_cache(conn, "test_app_test", {[id: user.id], []}, backend: ets)
+    id   = store_in_cache(conn, "test_app_test", {user, []})
     conn =
       conn
       |> persistent_cookie("test_app_" <> @cookie_key, id)
@@ -157,12 +153,11 @@ defmodule PowPersistentSession.Plug.CookieTest do
     assert %{value: new_id, max_age: @max_age, path: "/"} = conn.resp_cookies["test_app_" <> @cookie_key]
     assert {:ok, decoded_id} = Plug.verify_token(conn, Atom.to_string(Cookie), new_id)
     assert String.starts_with?(decoded_id, "test_app")
-    assert get_from_cache(conn, new_id, backend: ets) == {[id: 1], []}
+    assert get_from_cache(conn, new_id) == {user, []}
   end
 
-  test "call/2 when user already assigned", %{conn: conn, ets: ets} do
-    user = %User{id: 1}
-    id   = store_in_cache(conn, "test", {[id: user.id], []}, backend: ets)
+  test "call/2 when user already assigned", %{conn: conn, user: user} do
+    id   = store_in_cache(conn, "test", {user, []})
     conn =
       conn
       |> persistent_cookie(@cookie_key, id)
@@ -170,12 +165,12 @@ defmodule PowPersistentSession.Plug.CookieTest do
       |> run_plug()
 
     refute conn.resp_cookies[@cookie_key]
-    assert get_from_cache(conn, id, backend: ets) == {[id: 1], []}
+    assert get_from_cache(conn, id) == {user, []}
   end
 
-  test "call/2 when user doesn't exist in database", %{conn: conn, ets: ets} do
+  test "call/2 when user doesn't exist in database", %{conn: conn} do
     user = %User{id: -1}
-    id   = store_in_cache(conn, "test", {[id: user.id], []}, backend: ets)
+    id   = store_in_cache(conn, "test", {user, []})
     conn =
       conn
       |> persistent_cookie(@cookie_key, id)
@@ -183,13 +178,12 @@ defmodule PowPersistentSession.Plug.CookieTest do
 
     refute Plug.current_user(conn)
     refute conn.resp_cookies[@cookie_key]
-    assert get_from_cache(conn, id, backend: ets) == :not_found
+    assert get_from_cache(conn, id) == :not_found
   end
 
-  test "call/2 with unsigned token", %{conn: conn, ets: ets} do
-    user = %User{id: 1}
+  test "call/2 with unsigned token", %{conn: conn, user: user} do
     id = "test"
-    store_in_cache(conn, id, {[id: user.id], []}, backend: ets)
+    store_in_cache(conn, id, {user, []})
     conn =
       conn
       |> persistent_cookie(@cookie_key, id)
@@ -197,7 +191,8 @@ defmodule PowPersistentSession.Plug.CookieTest do
 
     refute Plug.current_user(conn)
     refute conn.resp_cookies[@cookie_key]
-    assert PersistentSessionCache.get([backend: ets], id) == {[id: user.id], []}
+
+    assert get_unsigned_from_cache(id) == {user, []}
   end
 
   test "call/2 when persistent session cache doesn't have credentials", %{conn: conn} do
@@ -210,21 +205,20 @@ defmodule PowPersistentSession.Plug.CookieTest do
     refute conn.resp_cookies[@cookie_key]
   end
 
-  test "call/2 with invalid stored clauses", %{conn: conn, ets: ets} do
-    user = %User{id: 1}
-    id   = store_in_cache(conn, "test", {[id: user.id, uid: 2], []}, backend: ets)
+  test "call/2 handles clause error", %{conn: conn} do
+    user = %User{id: nil}
+    id   = store_in_cache(conn, "test", {user, []})
 
-    assert_raise RuntimeError, "Invalid get_by clauses stored: [id: 1, uid: 2]", fn ->
+    assert_raise RuntimeError, "Primary key value for key `:id` in #{inspect User} can't be `nil`", fn ->
       conn
       |> persistent_cookie(@cookie_key, id)
       |> run_plug()
     end
-  end
+   end
 
   # TODO: Remove by 1.1.0
-  test "call/2 is backwards-compatible with just user fetch clause", %{conn: conn, ets: ets} do
-    user = %User{id: 1}
-    id   = store_in_cache(conn, "test", [id: user.id], backend: ets)
+  test "call/2 is backwards-compatible with user fetch clause", %{conn: conn, user: user} do
+    id   = store_in_cache(conn, "test", {[id: user.id], []})
     conn =
       conn
       |> persistent_cookie(@cookie_key, id)
@@ -233,14 +227,13 @@ defmodule PowPersistentSession.Plug.CookieTest do
     assert Plug.current_user(conn) == user
     assert %{value: new_id, max_age: @max_age, path: "/"} = conn.resp_cookies[@cookie_key]
     refute new_id == id
-    assert get_from_cache(conn, id, backend: ets) == :not_found
-    assert get_from_cache(conn, new_id, backend: ets) == {[id: 1], []}
+    assert get_from_cache(conn, id) == :not_found
+    assert get_from_cache(conn, new_id) == {user, []}
   end
 
   # TODO: Remove by 1.1.0
-  test "call/2 is backwards-compatible with `:session_fingerprint` metadata", %{conn: conn, ets: ets} do
-    user = %User{id: 1}
-    id   = store_in_cache(conn, "test", {[id: user.id], session_fingerprint: "fingerprint"}, backend: ets)
+  test "call/2 is backwards-compatible with just user fetch clause", %{conn: conn, user: user} do
+    id   = store_in_cache(conn, "test", [id: user.id])
     conn =
       conn
       |> persistent_cookie(@cookie_key, id)
@@ -249,17 +242,32 @@ defmodule PowPersistentSession.Plug.CookieTest do
     assert Plug.current_user(conn) == user
     assert %{value: new_id, max_age: @max_age, path: "/"} = conn.resp_cookies[@cookie_key]
     refute new_id == id
-    assert get_from_cache(conn, id, backend: ets) == :not_found
-    assert get_from_cache(conn, new_id, backend: ets) == {[id: 1], session_metadata: [fingerprint: "fingerprint"]}
+    assert get_from_cache(conn, id) == :not_found
+    assert get_from_cache(conn, new_id) == {user, []}
+  end
+
+  # TODO: Remove by 1.1.0
+  test "call/2 is backwards-compatible with `:session_fingerprint` metadata", %{conn: conn, user: user} do
+    id   = store_in_cache(conn, "test", {user, session_fingerprint: "fingerprint"})
+    conn =
+      conn
+      |> persistent_cookie(@cookie_key, id)
+      |> run_plug()
+
+    assert Plug.current_user(conn) == user
+    assert %{value: new_id, max_age: @max_age, path: "/"} = conn.resp_cookies[@cookie_key]
+    refute new_id == id
+    assert get_from_cache(conn, id) == :not_found
+    assert get_from_cache(conn, new_id) == {user, session_metadata: [fingerprint: "fingerprint"]}
     assert conn.private[:pow_session_metadata][:fingerprint] == "fingerprint"
   end
 
-  test "create/3 with custom TTL", %{conn: conn, config: config} do
+  test "create/3 with custom TTL", %{conn: conn, config: config, user: user} do
     config = Keyword.put(config, :persistent_session_ttl, 1000)
     conn   =
       conn
       |> init_plug(config)
-      |> run_create(%User{id: 1}, config)
+      |> run_create(user, config)
 
     assert_received {:ets, :put, [{_key, _value} | _rest], config}
     assert config[:ttl] == 1000
@@ -267,20 +275,12 @@ defmodule PowPersistentSession.Plug.CookieTest do
     assert %{max_age: 1, path: "/"} = conn.resp_cookies[@cookie_key]
   end
 
-  test "create/3 handles clause error", %{conn: conn, config: config} do
-    assert_raise RuntimeError, "Primary key value for key `:id` in #{inspect User} can't be `nil`", fn ->
-      conn
-      |> init_plug(config)
-      |> run_create(%User{id: nil}, config)
-    end
-  end
-
-  test "create/3 with custom cookie options", %{conn: conn, config: config} do
+  test "create/3 with custom cookie options", %{conn: conn, config: config, user: user} do
     config = Keyword.put(config, :persistent_session_cookie_opts, @custom_cookie_opts)
     conn   =
       conn
       |> init_plug(config)
-      |> run_create(%User{id: 1}, config)
+      |> run_create(user, config)
 
     assert %{
       domain: "domain.com",
@@ -292,40 +292,40 @@ defmodule PowPersistentSession.Plug.CookieTest do
     } = conn.resp_cookies[@cookie_key]
   end
 
-  test "create/3 deletes previous persistent session", %{conn: conn, config: config, ets: ets} do
-    id   = store_in_cache(conn, "previous_persistent_session", {[id: 1], []}, backend: ets)
+  test "create/3 deletes previous persistent session", %{conn: conn, config: config, user: user} do
+    id   = store_in_cache(conn, "previous_persistent_session", {user, []})
     conn = persistent_cookie(conn, @cookie_key, id)
 
-    assert get_from_cache(conn, id, backend: ets) == {[id: 1], []}
+    assert get_from_cache(conn, id) == {user, []}
 
     conn
     |> init_plug(config)
-    |> run_create(%User{id: 1}, config)
+    |> run_create(user, config)
 
-    assert_received {:ets, :put, [{_key, {[id: 1], []}}], _config}
-    assert get_from_cache(conn, id, backend: ets) == :not_found
+    assert_received {:ets, :put, [{_key, {^user, []}}], _config}
+    assert get_from_cache(conn, id) == :not_found
   end
 
-  test "create/3 with `[:pow_session_metadata][:fingerprint]` defined in conn.private", %{conn: conn, config: config} do
+  test "create/3 with `[:pow_session_metadata][:fingerprint]` defined in conn.private", %{conn: conn, config: config, user: user} do
     conn
     |> Conn.put_private(:pow_session_metadata, fingerprint: "fingerprint")
     |> init_plug(config)
-    |> run_create(%User{id: 1}, config)
+    |> run_create(user, config)
 
-    assert_received {:ets, :put, [{_key, {[id: 1], session_metadata: [fingerprint: "fingerprint"]}}], _config}
+    assert_received {:ets, :put, [{_key, {user, session_metadata: [fingerprint: "fingerprint"]}}], _config}
   end
 
-  test "create/3 with custom metadata", %{conn: conn, config: config} do
+  test "create/3 with custom metadata", %{conn: conn, config: config, user: user} do
     conn
     |> Conn.put_private(:pow_persistent_session_metadata, session_metadata: [a: 1])
     |> init_plug(config)
-    |> run_create(%User{id: 1}, config)
+    |> run_create(user, config)
 
-    assert_received {:ets, :put, [{_key, {[id: 1], session_metadata: [a: 1]}}], _config}
+    assert_received {:ets, :put, [{_key, {user, session_metadata: [a: 1]}}], _config}
   end
 
-  test "delete/3", %{conn: conn, ets: ets, config: config} do
-    id   = store_in_cache(conn, "test", {[id: 1], []}, backend: ets)
+  test "delete/3", %{conn: conn, config: config, user: user} do
+    id   = store_in_cache(conn, "test", {user, []})
     conn =
       conn
       |> persistent_cookie(@cookie_key, id)
@@ -333,11 +333,11 @@ defmodule PowPersistentSession.Plug.CookieTest do
       |> run_delete(config)
 
     assert conn.resp_cookies[@cookie_key] == %{max_age: 0, path: "/", universal_time: {{1970, 1, 1}, {0, 0, 0}}}
-    assert get_from_cache(conn, id, backend: ets) == :not_found
+    assert get_from_cache(conn, id) == :not_found
   end
 
-  test "delete/3 with custom cookie options", %{conn: conn, ets: ets, config: config} do
-    id     = store_in_cache(conn, "test", {[id: 1], []}, backend: ets)
+  test "delete/3 with custom cookie options", %{conn: conn, config: config, user: user} do
+    id     = store_in_cache(conn, "test", {user, []})
     config = Keyword.put(config, :persistent_session_cookie_opts, @custom_cookie_opts)
     conn   =
       conn
@@ -346,7 +346,7 @@ defmodule PowPersistentSession.Plug.CookieTest do
       |> run_delete(config)
 
     assert conn.resp_cookies[@cookie_key] == %{max_age: 0, universal_time: {{1970, 1, 1}, {0, 0, 0}}, path: "/path", domain: "domain.com", extra: "SameSite=Lax", http_only: false, secure: true}
-    assert get_from_cache(conn, id, backend: ets) == :not_found
+    assert get_from_cache(conn, id) == :not_found
   end
 
   defp conn_with_session_plug(config) do
@@ -383,15 +383,23 @@ defmodule PowPersistentSession.Plug.CookieTest do
     |> Conn.send_resp(200, "")
   end
 
-  defp store_in_cache(conn, token, value, config) do
-    PersistentSessionCache.put(config, token, value)
+  defp store_in_cache(conn, token, value) do
+    {store, opts} = Base.store(conn.private[:pow_config])
+
+    store.put(opts, token, value)
 
     Plug.sign_token(conn, Atom.to_string(Cookie), token)
   end
 
-  defp get_from_cache(conn, token, config) do
+  defp get_from_cache(conn, token) do
     assert {:ok, token} = Plug.verify_token(conn, Atom.to_string(Cookie), token)
 
-    PersistentSessionCache.get(config, token)
+    get_unsigned_from_cache(token)
+  end
+
+  defp get_unsigned_from_cache(token) do
+    {store, opts} = Base.store(PowPersistentSession.Test.pow_config())
+
+    store.get(opts, token)
   end
 end