+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v1.0.22 (TBA)
+
+### Enhancements
+
+* [`PowPersistentSession.Plug.Cookie`] Now stores the user struct instead of clauses
+* [`PowPersistentSession.Plug.Base`] Now includes `:pow_config` in the store config
+* [`PowResetPassword.Plug`] Now includes `:pow_config` in the store config
+* [`Pow.Plug.Base`] Now includes `:pow_config` in the store config
+* [`Pow.Operations`] Added `Pow.Operations.reload/2` to reload structs
+* [`PowPersistentSession.Store.PersistentSessionCache`] Update `PowPersistentSession.Store.PersistentSessionCache.get/2` to reload the user using `Pow.Operations.reload/2`
+* [`Pow.Store.CredentialsCache`] Now support `reload: true` configuration so once fetched from the cache the user object will be reloaded through the context module
+
 ## v1.0.21 (2020-09-13)
 
 ### Enhancements